Refactor in multiple files + stub for column visibility UI
